🚚 Rename:  editColorStyle -> color 파일명 변경
