+++ src/utils/colorLuminance.ts
@@ -1,6 +1,6 @@
 import { ColorType, CSSColorKeywords, HEX, RGB, RGBA } from '@/types/color';
 import { CSS_COLOR_KEYWORDS } from '@/constants/color';
-import { editColor } from './editColorStyle';
+import { editColor } from './color';
 
 type RgbObject = { r: number; g: number; b: number };
